Did more refactoring and I created a new character_class.py to contain all the code for the player class and enemy class instead of keeping those in two separate files. There's also a bug I need to fix when Dante attacks the enemy and defeats it. The fight continues even though the enemy is dead.
